feat: create account settings page
