+++ web/ui/src/containers/settings/SettingsLayout.tsx
@@ -25,8 +25,15 @@ export const SettingsLayout: NextPage<
               <MenuItem
                 active={page === 'profile'}
                 href="/settings/profile"
-                icon="fa-user"
+                icon="fa-id-badge"
                 name="Public profile"
+                rightChildren={<NewBadgy />}
+              />
+              <MenuItem
+                active={page === 'account'}
+                href="/settings/account"
+                icon="fa-user"
+                name="Account"
                 rightChildren={<NewBadgy />}
               />
               <MenuItem
@@ -42,15 +49,6 @@ export const SettingsLayout: NextPage<
                 name="Awesomess"
                 rightChildren={<NewBadgy />}
                 className='!text-fuchsia-400 dark:!text-fuchsia-600 hover:!bg-fuchsia-500/20 hover:!text-fuchsia-500 [&_.anchor-sub-text]:hover:!text-fuchsia-500 [&[data-active="true"]]:!bg-fuchsia-500/20 [&[data-active="true"]_.anchor-sub-text]:!text-fuchsia-500 '
-              />
-            </MenuCategory>
-            <MenuCategory name="Access">
-              <MenuItem
-                active={page === 'accounts'}
-                href="/settings/accounts"
-                icon="fa-shield-check"
-                name="Accounts"
-                rightChildren={<NewBadgy />}
               />
             </MenuCategory>
             <MenuCategory name="About">
